+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface UserService {
     List<User> getAllUsers();
 
-    User getUserById(int id);
+    User getUserById(long id);
 
     User addUser(User user);
 
-    User updateUser(int id, User user);
+    User updateUser(long id, User user);
 
-    boolean deleteUser(int id);
+    boolean deleteUser(long id);
 }
